# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "CIMOL API with Golang.\nGolang memiliki keunggulan dalam skalabilitas horizontal dan pemeliharaan kode yang lebih sederhana,\nGolang juga memiliki keamanan bawaan dan memudahkan pencegahan beberapa jenis serangan, serta\nkinerja yang signifikan dalam kasus pengolahan beban kerja tinggi dan pembuatan API yang membutuhkan respons cepat.",
        "title": "CIMOL API with Golang",
        "contact": {
            "name": "API CNAF MOBILE",
            "url": "https://www.cnaf.co.id/",
            "email": "dimas.zikri@cnaf.co.id"
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/api",
    "paths": {
...
        "/tools/Me": {
            "get": {
                "description": "OAuth protects our entity endpoints",
                "tags": [
                    "MyProfile"
                ],
                "summary": "Get Data Assign FC",
                "parameters": [
                    {
                        "type": "string",
                        "description": "string example",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.MyProfile"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseWithoutData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "helpers.ResponseWithData": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "responsecode": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "helpers.ResponseWithoutData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "responsecode": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.MyProfile": {
            "type": "object",
            "properties": {
                "real_name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}